+++ src/main/java/com/nhnacademy/store99/bookstore/book/repository/impl/BookRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.store99.bookstore.book.repository.impl;
 
-import com.nhnacademy.store99.bookstore.book.dto.response.response.BookResponse;
+import com.nhnacademy.store99.bookstore.book.dto.response.BookResponse;
 import com.nhnacademy.store99.bookstore.book.entity.Book;
 import com.nhnacademy.store99.bookstore.book.entity.QBook;
 import com.nhnacademy.store99.bookstore.book.repository.BookRepository;
