+++ src/main/java/com/nhnacademy/store99/bookstore/book/service/BookService.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.store99.bookstore.book.service;
 
+import com.nhnacademy.store99.bookstore.book.dto.response.BookResponse;
 import com.nhnacademy.store99.bookstore.book.dto.response.SimpleBookResponse;
-import com.nhnacademy.store99.bookstore.book.dto.response.response.BookResponse;
 import java.util.List;
 import java.util.Set;
