+++ src/main/java/com/nhnacademy/store99/bookstore/book_category/service/BookCategoryService.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.store99.bookstore.book_category.service;
 
-import com.nhnacademy.store99.bookstore.book.dto.response.response.BookListElementDTO;
+import com.nhnacademy.store99.bookstore.book.dto.response.BookListElementDTO;
 import com.nhnacademy.store99.bookstore.book_category.dto.response.CategoryParentsDTO;
 import java.util.List;
 import org.springframework.data.domain.Page;
